requirements.txt
streamlit
sentence_transformers
scikit-learn
tiktoken
pandas
openpyxl
umap-learn
numpy
meilisearch
pylette
botocore
boto3
python-dotenv
requests
pillow